Split index jsx file based on component
