+++ src/components/button.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Button({ children, onClick }) {
+  return (
+    <button type="button" onClick={() => onClick(children)}>
+      {children}
+    </button>
+  );
+}
+
+export default Button;
