+++ src/index.jsx
@@ -1,52 +1,6 @@
 import React, { useState } from 'react';
 import ReactDOM from 'react-dom';
-
-function Counter({ count, onClick }) {
-  return (
-    <button type="button" onClick={onClick}>
-      Click me!
-      (
-      { count }
-      )
-    </button>
-  );
-}
-
-function Buttons({ onClick }) {
-  return (
-    <p>
-      {[1, 2, 3, 4, 5].map((i) => (
-        <Button key={i} onClick={onClick}>
-          {i}
-        </Button>
-      ))}
-    </p>
-  );
-}
-
-function Button({ children, onClick }) {
-  return (
-    <button type="button" onClick={() => onClick(children)}>
-      {children}
-    </button>
-  );
-}
-
-function Page({ count, onClick }) {
-  return (
-    <div>
-      <p>Hello, world!!!</p>
-      <p>Hi!</p>
-      <Counter
-        count={count}
-        onClick={onClick}
-      />
-      <Buttons
-        onClick={onClick}
-      />
-    </div>
-  );
-}
+import Page from './pages/page';
 
 function App() {
   const [state, setState] = useState({
